Remove LoginView
Redundant, especially with incoming improvements to the AddServerView
