+++ Stingray/Models/RErrorModel.swift
@@ -97,6 +97,43 @@ public enum NetworkError: RError {
             return "An access token is needed"
         }
     }
+    
+    /// A function to override `NetworkError` messages with a more human readable option
+    /// - Parameters:
+    ///   - netErr: NetworkError that was thrown
+    ///   - httpProtocol: HTTP protocol used
+    /// - Returns: Formatted error
+    static func overrideNetErrorMessage(netErr: NetworkError, httpProtocol: HttpProtocol) -> String {
+        switch netErr {
+        case .invalidURL:
+            switch httpProtocol {
+            case .http: return "Invalid HTTP URL. Check your hostname and port."
+            case .https: return "Invalid HTTPS URL. Check your URL."
+            }
+        case .encodeJSONFailed: return "Failed to send request to server. " +
+                "This may be because of some tricky characters in your username and password."
+        case .decodeJSONFailed, .missingAccessToken, .requestFailedToSend:
+            switch httpProtocol {
+            case .http: return "Could not find your Jellyfin server. Please check your hostname and port."
+            case .https: return "Could not find your Jellyfin server. Please check your URL."
+            }
+        case .badResponse(let responseCode, _):
+            switch responseCode {
+            case 401: return "Invalid username or password."
+            case 404:
+                switch httpProtocol {
+                case .http: return "Could not find your Jellyfin server. Please check your hostname and port."
+                case .https: return "Could not find your Jellyfin server. Please check your URL."
+                }
+            default: return "An unexpected error occurred. Please make sure your login details are correct."
+            }
+        }
+    }
+}
+
+enum HttpProtocol: String, CaseIterable {
+    case http = "http"
+    case https = "https"
 }
 
 /// Different ways JSON can have an error.
